fix test and introduce option to truncate on schema change

+++ src/spetlr/deltaspec/DeltaTableSpecDifference.py
@@ -174,9 +174,17 @@ class DeltaTableSpecDifference:
         allow_columns_type_change=False,
         errors_as_warnings=False,
         allow_columns_reorder=False,
+        schema_change_as_truncate=False,
     ) -> List[str]:
         if self.base.schema == self.target.schema:
             return []
+
+        allow_columns_add = allow_columns_add or schema_change_as_truncate
+        allow_columns_drop = allow_columns_drop or schema_change_as_truncate
+        allow_columns_type_change = (
+            allow_columns_type_change or schema_change_as_truncate
+        )
+        allow_columns_reorder = allow_columns_reorder or schema_change_as_truncate
 
         statements = []
         base_fields: Dict[str, StructField] = {
@@ -251,7 +259,7 @@ class DeltaTableSpecDifference:
                 fields_to_remove.append(key)
 
         # Now convert the accumulated changes to ALTER statements.
-        if fields_to_remove:
+        if fields_to_remove and not schema_change_as_truncate:
             statement = f"ALTER TABLE {self.base.name} DROP COLUMN"
             if len(fields_to_remove) > 1:
                 statement += "S"
@@ -259,7 +267,7 @@ class DeltaTableSpecDifference:
 
             statements.append(statement)
 
-        if fields_to_add:
+        if fields_to_add and not schema_change_as_truncate:
             statement = f"ALTER TABLE {self.base.name} ADD COLUMN"
             if len(fields_to_add) == 1:
                 statement += (
@@ -291,6 +299,13 @@ class DeltaTableSpecDifference:
                 f"ALTER TABLE {self.base.name} ALTER COLUMN "
                 f"{key} COMMENT {json.dumps(newComment)}"
             )
+
+        if schema_change_as_truncate and (fields_to_add or fields_to_remove):
+            statements.append(f"TRUNCATE TABLE {self.base.name}")
+            # If the truncate option is chosen, we don't need to reorder.
+            # That will be taken care of by the overwriteSchema operation
+            # That we expect will follow.
+            return statements
 
         # Finally address the possible change of order.
         base_order_so_far = copy.copy(base_order)
@@ -333,7 +348,18 @@ class DeltaTableSpecDifference:
 
         return statements
 
-    def _alter_statements_for_new_location(self, allow_columns_add=False) -> List[str]:
+    def _alter_statements_for_new_location(
+        self,
+        allow_columns_add=False,
+        allow_columns_drop=False,
+        allow_columns_type_change=False,
+        allow_columns_reorder=False,
+        allow_name_change=False,
+        allow_location_change=False,
+        allow_table_create=True,
+        errors_as_warnings=False,
+        schema_change_as_truncate=False,
+    ) -> List[str]:
         """In case we need to change location,
         the alter statements will consist of creating the new target.
         and then pointing the current table at it. The target will be empty."""
@@ -358,11 +384,22 @@ class DeltaTableSpecDifference:
 
         statements += DeltaTableSpecDifference(
             target=self.target, base=new_base
-        ).alter_statements(allow_columns_add=allow_columns_add)
+        ).alter_statements(
+            allow_columns_add=allow_columns_add,
+            allow_columns_drop=allow_columns_drop,
+            allow_columns_type_change=allow_columns_type_change,
+            allow_columns_reorder=allow_columns_reorder,
+            allow_name_change=allow_name_change,
+            allow_location_change=allow_location_change,
+            allow_table_create=allow_table_create,
+            errors_as_warnings=errors_as_warnings,
+            schema_change_as_truncate=schema_change_as_truncate,
+        )
         return statements
 
     def alter_statements(
         self,
+        *,
         allow_columns_add=False,
         allow_columns_drop=False,
         allow_columns_type_change=False,
@@ -371,6 +408,7 @@ class DeltaTableSpecDifference:
         allow_location_change=False,
         allow_table_create=True,
         errors_as_warnings=False,
+        schema_change_as_truncate=False,
     ) -> List[str]:
         """A list to alter statements that will ensure
         that what used to be the base table, becomes the target table"""
@@ -410,6 +448,7 @@ class DeltaTableSpecDifference:
                     allow_location_change=allow_location_change,
                     allow_table_create=allow_table_create,
                     errors_as_warnings=errors_as_warnings,
+                    schema_change_as_truncate=schema_change_as_truncate,
                 )
             else:
                 print(
@@ -423,7 +462,15 @@ class DeltaTableSpecDifference:
         if base.location != target.location:
             if allow_location_change:
                 return full_diff._alter_statements_for_new_location(
-                    allow_columns_add=allow_columns_add
+                    allow_columns_add=allow_columns_add,
+                    allow_columns_drop=allow_columns_drop,
+                    allow_columns_type_change=allow_columns_type_change,
+                    allow_columns_reorder=allow_columns_reorder,
+                    allow_name_change=allow_name_change,
+                    allow_location_change=allow_location_change,
+                    allow_table_create=allow_table_create,
+                    errors_as_warnings=errors_as_warnings,
+                    schema_change_as_truncate=schema_change_as_truncate,
                 )
             else:
                 if errors_as_warnings:
@@ -445,6 +492,7 @@ class DeltaTableSpecDifference:
             allow_columns_type_change=allow_columns_type_change,
             allow_columns_reorder=allow_columns_reorder,
             errors_as_warnings=errors_as_warnings,
+            schema_change_as_truncate=schema_change_as_truncate,
         )
 
         if base.comment != target.comment:
